Optimize build:clean task

+++ lib/antwort/tasks/build.rake
@@ -13,11 +13,11 @@ namespace :build do
     build_dir = File.expand_path('./build')
     next unless File.directory?(build_dir)
 
-    Dir.entries(build_dir)
-      .delete_if { |f| f.to_s[0] == '.' }
-      .each do |f|
-        puts "Removing #{f}..."
-        FileUtils.rm_rf(File.expand_path("./build/#{f}"))
-      end
+    Dir.foreach(build_dir) do |f|
+      next if f.to_s[0] == '.'
+
+      puts "Removing #{f}..."
+      FileUtils.rm_rf(File.expand_path("./build/#{f}"))
+    end
   end
 end
